refactor: modularized screens, added attendance with stable employee flow

+++ app/src/main/java/com/employeedb/employeedatabase/model/Employee.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "employees")
 data class Employee(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
     val email: String,
     val role: String,
